factorisation makePathFromTip dans tree-updates.js
